Refactor to split visit details page
